// src/models/User.ts
export interface User {
    id: string;
    firstName: string;
    lastName: string;
    gender: string;
    email: string;
    password: string;
    birthday: Date;
    telephone: string;
    country: string;
    bio: string;
    favoriteNumber: string;
    favoriteColor: string;
    avatarImagePath: string;
    agreementLevel: string;
    getsNewsletter: boolean;
    createdAt: Date;
}